refactor(your-story): abstract localStorage login into utility functions
