+++ src/app/(pages)/fathers-and-figures/YourStory/MyStoriesProvider.tsx
@@ -1,11 +1,11 @@
 import React, { createContext, useEffect, useContext, useState } from 'react'
 import { StoryObjectType } from '../types'
+import { getLocalStories, setLocalStories } from './handleLocalStories'
 
 interface MyStoriesContextType {
   myStories: StoryObjectType[];
   setMyStories: React.Dispatch<React.SetStateAction<StoryObjectType[]>>
 }
-
 
 export const MyStoriesContext = createContext< MyStoriesContextType | undefined >(undefined)
 
@@ -14,26 +14,11 @@ export const MyStoriesProvider = ({ children }: { children: React.ReactNode }) =
   const [myStories, setMyStories] = useState<StoryObjectType[]>(() => [])
   
   useEffect(() => {
-    const localStories = localStorage.getItem('localStories')
-    if (localStories) {
-      try {
-        const parsedStories = JSON.parse(localStories)
-        if (!Array.isArray(parsedStories)) {
-          throw new Error('Invalid local stories format')
-        }
-        setMyStories(parsedStories)
-      } catch (error){
-        console.error('Failed to parse local stories:', error)
-      }
-    }
+    setMyStories(getLocalStories())
   }, [])
   
   useEffect(() => {
-    try {
-      localStorage.setItem('localStories', JSON.stringify(myStories))
-    } catch (error){
-      console.error('Failed to save local stories:', error)
-    }
+    setLocalStories(myStories)
   }, [myStories])
 
   return (
