--- src/app/(pages)/fathers-and-figures/YourStory/handleMyStories.ts
@@ -1,1 +1,0 @@
-
